+++ src/Api/Strategies/xWing.ts
@@ -1,8 +1,7 @@
 import { ALL_CANDIDATES, IndexToNine, INDICES_TO_NINE, SudokuDigits } from "../../Types";
 import PureSudoku from "../Spaces/PureSudoku"
 import Solver from "../Solver";
-import { CellID } from "../Utils";
-import { colorGroup } from "./intersectionRemoval";
+import { CellID, colorGroup } from "../Utils";
 
 export function _innerWingLogic(
    candidate: SudokuDigits,